\name{logNlogS.broken-package}
\alias{logNlogS.broken-package}
\docType{package}
\title{
Analysis of \emph{logN-logS} relationship with broken-power laws
}
\description{
This package contains some routines for conducting analysis on the relationship between the number of astronomical sources and their flux. This relationship is typically known as \emph{log(N) - log(S)} relationship. This package provides maximum likelihood estimation on the typical \emph{log(N) - log(S)} model and the broken power-law models.
}
\details{
\tabular{ll}{
Package: \tab logNlogS\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-03-28\cr
License: \tab What license is it under?\cr
}

}
\author{
Raymond K. W. Wong

Maintainer: Raymond K. W. Wong <raymondkww.dev@gmail.com>
}
%\references{}

\keyword{ package }
%\seealso{ }
\examples{
# TODO ~~ simple examples of the most important functions ~~
}
